SCRS-7684 Removed holding pen and timepoint functionality
