+++ conf/application.conf
@@ -167,12 +167,6 @@ microservice {
       port = 9982
     }
 
-    coho-api {
-      url = "https://ewfgonzo.companieshouse.gov.uk/submissions"
-      stub-url = "http://localhost:9986/incorporation-frontend-stubs/submissions"
-      token = "Fak3-t0k3N-f0R-C0H0-APi-iN-l0CaL-MaCHiN3"
-    }
-
     gov-uk {
       register-your-company = "https://www.gov.uk/limited-company-formation/register-your-company"
     }
