+++ it/itutil/FakeAppConfig.scala
@@ -36,10 +36,7 @@ trait FakeAppConfig {
     "microservice.services.company-registration.port" -> s"$mockPort",
     "microservice.services.business-registration.host" -> s"$mockHost",
     "microservice.services.business-registration.port" -> s"$mockPort",
-    "microservice.services.coho-api.stub-url" -> s"$mockUrl/check-submission",
-    "microservice.services.coho-api.url" -> mockUrl,
-    "microservice.timeoutInSeconds" -> "999999",
-    "microservice.services.coho-api.token" -> "abc"
+    "microservice.timeoutInSeconds" -> "999999"
   ) ++ extraConfig
 
 }
